dosya düzenlemesi
