+++ src/components/header-components/Dropdown.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
-import { getCategories } from "../store/actions/globalAction";
-import { getProducts } from "../store/actions/productAction";
+import { getCategories } from "../../store/actions/globalAction";
+import { getProducts } from "../../store/actions/productAction";
 
 const TestDropdown = () => {
   const [isOpen, setIsOpen] = useState(false);
