+++ src/layouts/Header.jsx
@@ -1,21 +1,14 @@
-import NavbarDark from "../components/NavbarDark";
-import NavbarLight from "../components/NavbarLight";
-import NavbarLight2 from "../components/NavbarLight2";
+import NavbarDark from "../components/header-components/NavbarDark";
+import NavbarLight from "../components/header-components/NavbarLight";
+
 import { useLocation } from "react-router-dom";
 import React, { useState, useEffect } from "react";
 
 const Header = () => {
-  const { pathname } = useLocation();
-
-  /*collapse={collapse}
-        setCollapse={setCollapse}
-        userCollapse={userCollapse}
-        setUserCollapse={setUserCollapse} */
   return (
     <header className="w-full  font-Montserrat text-white">
       <NavbarDark />
       <NavbarLight />
-
       <hr />
     </header>
   );
