+++ src/pages/HomePage.jsx
@@ -1,10 +1,12 @@
-import Body from "../layouts/Body";
-import Slider from "../components/Slider";
+import Slider from "../components/homepage-components/Slider";
 import { Link, useLocation } from "react-router-dom";
 import { useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
 import { getProducts } from "../store/actions/productAction";
 import Loading from "../components/Loading";
+import { ShopCards } from "../components/homepage-components/ShopCards";
+import { ShopCards2 } from "../components/homepage-components/ShopCards2";
+import { Blog } from "../components/homepage-components/Blog";
 
 const HomePage = () => {
   const [loading, setLoading] = useState(true);
@@ -28,7 +30,9 @@ const HomePage = () => {
       ) : (
         <div className="flex flex-col gap-2">
           <Slider />
-          <Body />
+          <ShopCards />
+          <ShopCards2 />
+          <Blog />
         </div>
       )}
     </>
