--- src/pages/ProductListPages.jsx
@@ -1,10 +1,0 @@
-import { Shop } from "../components/Shop";
-
-const ProductListPages = () => {
-  return (
-    <div>
-      <Shop />
-    </div>
-  );
-};
-export default ProductListPages;
